feat: [ON-1192] private and public homepage
