+++ app/src/app/[lng]/[inventory]/InventoryResultTab/index.tsx
@@ -17,7 +17,7 @@ import {
   Tabs,
   Text,
 } from "@chakra-ui/react";
-import { TabHeader } from "@/app/[lng]/[inventory]/TabHeader";
+import { TabHeader } from "@/components/HomePage/TabHeader";
 import EmissionsWidget from "@/app/[lng]/[inventory]/InventoryResultTab/EmissionsWidget";
 import TopEmissionsWidget from "@/app/[lng]/[inventory]/InventoryResultTab/TopEmissionsWidget";
 import { BlueSubtitle } from "@/components/blue-subtitle";
