+++ app/src/app/[lng]/public/[inventory]/page.tsx
@@ -1,11 +1,9 @@
-"use client";
-
 import HomePage from "@/components/HomePage/HomePage";
 
-export default function PrivateHome({
+export default function PublicHome({
   params: { lng },
 }: {
   params: { lng: string };
 }) {
-  return <HomePage lng={lng} isPublic={false} />;
+  return <HomePage lng={lng} isPublic={true} />;
 }
